+++ oddt/toolkits/rdk.py
@@ -421,13 +421,13 @@ class Molecule(object):
         # Atoms
         atom_dtype = [('id', 'int16'),
                  # atom info
-                 ('coords', 'float16', 3),
-                 ('radius', 'float16'),
-                 ('charge', 'float16'),
+                 ('coords', 'float32', 3),
+                 ('radius', 'float32'),
+                 ('charge', 'float32'),
                  ('atomicnum', 'int8'),
                  ('atomtype','a4'),
                  ('hybridization', 'int8'),
-                 ('neighbors', 'float16', (4,3)), # non-H neighbors coordinates for angles (max of 6 neighbors should be enough)
+                 ('neighbors', 'float32', (4,3)), # non-H neighbors coordinates for angles (max of 6 neighbors should be enough)
                  # residue info
                  ('resid', 'int16'),
                  ('resname', 'a3'),
@@ -464,7 +464,7 @@ class Molecule(object):
             residue = None
 
             # get neighbors, but only for those atoms which realy need them
-            neighbors = np.zeros(4, dtype=[('coords', 'float16', 3),('atomicnum', 'int8')])
+            neighbors = np.zeros(4, dtype=[('coords', 'float32', 3),('atomicnum', 'int8')])
             neighbors['coords'].fill(np.nan)
             for n, nbr_atom in enumerate(atom.neighbors):
                 neighbors[n] = (nbr_atom.coords, nbr_atom.atomicnum)
@@ -519,9 +519,9 @@ class Molecule(object):
 #            # Protein Residues (alpha helix and beta sheet)
 #            res_dtype = [('id', 'int16'),
 #                         ('resname', 'a3'),
-#                         ('N', 'float16', 3),
-#                         ('CA', 'float16', 3),
-#                         ('C', 'float16', 3),
+#                         ('N', 'float32', 3),
+#                         ('CA', 'float32', 3),
+#                         ('C', 'float32', 3),
 #                         ('isalpha', 'bool'),
 #                         ('isbeta', 'bool')
 #                         ] # N, CA, C
@@ -567,7 +567,7 @@ class Molecule(object):
                 # get vector perpendicular to ring
                 vector = np.cross(coords - np.vstack((coords[1:],coords[:1])), np.vstack((coords[1:],coords[:1])) - np.vstack((coords[2:],coords[:2]))).mean(axis=0) - centroid
                 r.append((centroid, vector, atom['isalpha'], atom['isbeta']))
-        ring_dict = np.array(r, dtype=[('centroid', 'float16', 3),('vector', 'float16', 3),('isalpha', 'bool'),('isbeta', 'bool'),])
+        ring_dict = np.array(r, dtype=[('centroid', 'float32', 3),('vector', 'float32', 3),('isalpha', 'bool'),('isbeta', 'bool'),])
 
         self._atom_dict = atom_dict
         self._ring_dict = ring_dict
